requirements.txt
alembic==1.13.1
Authlib==1.3.0
blinker==1.7.0
cachelib==0.9.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
ecdsa==0.19.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-DotEnv==0.1.2
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.5
MarkupSafe==2.1.5
msgspec==0.18.6
pyasn1==0.6.0
pycparser==2.22
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
